+++ large_image/tilesource/utilities.py
@@ -649,9 +649,9 @@ def getPaletteColors(value: Union[str, List[Union[str, float, Tuple[float, ...]]
             if value in mpl.colors.get_named_colors_mapping():
                 palette = ['#0000', mpl.colors.to_hex(str(value))]
             else:
-                cmap = mpl.colormaps.get_cmap(str(value)) if hasattr(getattr(
-                    mpl, 'colormaps', None), 'get_cmap') else mpl.cm.get_cmap(
-                        str(value))
+                cmap = (mpl.colormaps.get_cmap(str(value)) if hasattr(getattr(
+                    mpl, 'colormaps', None), 'get_cmap') else
+                    mpl.cm.get_cmap(str(value)))  # type: ignore
                 palette = [mpl.colors.to_hex(cmap(i)) for i in range(cmap.N)]
         except (ImportError, ValueError, AttributeError):
             pass